Create subpackage container-base
This package installs the dependencies and files which are required on
the base TripleO container image.

Currently, this is the puppet package and the bootstrap_host_exec script.

This is proposed as an alternative to packaging bootstrap_host_exec in
openstack-tripleo-common in Ia2c82e67335b2806cbe870b6f4ccb3176c7458f4.

The container-base package is consumed upstream by
https://review.openstack.org/464094

Change-Id: Iae260ce34770e90b89d82b992461ea66f5fd2d5a