Drop instack-undercloud dependency
The instack-undercloud dependency was added because the image building
yaml files reference elements from instack-undercloud. But these image
yaml files are actually used by python-tripleoclient which already has
the instack-undercloud dependency.

Additionally we would like to add tripleo-common as a dependency for
tripleo-heat-templates because we are looking to use some of the python
bits in CI and for the end-user but instack-undercloud requires THT so
this also prevents any dependency cycles.

Change-Id: I49c1c00d2175690df859e6f5c868d516bd6fc919