Revert to python-docker-py for tripleo-common
The tripleo-common commit that introduced the new python-docker
dependency [1] was made in a backward compatible way with the
python-docker-py library.

The main motivation to revert this change is that most of the testing
in the pike cycle was done using python-docker-py until the missing
dependency was noticed and added right before the release of pike.

The python-docker package obsoletes python-docker-py, so nothing breaks
if it is already installed.

[1] https://github.com/openstack/tripleo-common/commit/49c5372d8c7f0c726fe6b27a1ac610154677d775

Change-Id: Ib2dab0f104dae2c3787a6fe33fa75ad4473e3df0