Depend on skopeo
As part of upstream bug 1708965 and bug 1694709, tranferring images
between registries and introspecting registry images will be performed
by skopeo instead of the docker client. The benefits of using skopeo
are:
- images are not stored in the local docker when transferring between
  registries
- registries can be queried for image metadata with out pulling them
  to the local docker
- proper support for valid protocol based URLs when referring to image
  registry locations

This dependency is added before any upstream changes are proposed so
that the package will not break when they are.

Change-Id: Ifdd0c52efbb69fee80a6db460142a1c90674cd5e