Make sure that we pull in all the needed rpms for the binaries used by healthechecks
The tripleo-common-contrainer-base package contains the healthchecks.
It's main file 'common.sh' uses the following binaries:
lsof, pgrep, sudo, crudini, curl, ss

We should make sure that we depend on the packges providing those
commands. We already do so for lsof and hostname, let's make sure we
pull in the others too. That way we do not pull them in via other means
or by sheer luck.

First observed via rhbz#1885520

Change-Id: Ifbd97755c7b6e948a0b9d16d9c107643421a28c2