[train-only] Allows to "correct" and fix the python version
A new script is being added via [1]. Since it's python, it has to
support both python2 and python3 because of the CentOS-7 and CentOS-8
support of this release.

If we just push "/usr/bin/python", it fails on the CentOS-8 content
provider job with the following:
*** ERROR: ambiguous python shebang in /usr/share/openstack-tripleo-common/healthcheck/http-healthcheck.py: #!/usr/bin/python. Change it to python3 (or python2) explicitly.

This patch attempts to fix the version during the package build.

[1] https://review.opendev.org/c/openstack/tripleo-common/+/772412

Change-Id: I2453a5baf5c2b3aa421853cd9734db95c2dbc4be